user_service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-jose==3.5.0
pydantic==2.11.5
pydantic[email]
passlib[bcrypt]==1.7.4
python-multipart